Refactor geometry helpers
