+++ generate_parking_slots.py
@@ -3,6 +3,7 @@ from shapely.geometry import LineString, Polygon, mapping, MultiPoint
 import json
 import math
 import numpy as np
+from geometry_utils import bearing, get_point_at_distance, haversine_distance
 
 reverse_coordinates = True
 
@@ -22,70 +23,7 @@ config_geometry = {}
 
 parking_slot_width = 0.003854667
 parking_slot_length = 0.014
-R  = 6371e3  # earth radius in meters
 
-# Converts from degrees to radians.
-def toRadians(degrees):
-    return degrees * math.pi / 180
-
-# Converts from radians to degrees.
-def toDegrees(radians):
-    return radians * 180 / math.pi
-
-def bearing(startLat, startLng, destLat, destLng):
-    """
-    Calculate the initial bearing using 2 points
-    returns bearing in degrees
-    """
-    startLat = toRadians(startLat)
-    startLng = toRadians(startLng)
-    destLat = toRadians(destLat)
-    destLng = toRadians(destLng)
-
-    y = math.sin(destLng - startLng) * math.cos(destLat)
-    x = math.cos(startLat) * math.sin(destLat) - math.sin(startLat) * math.cos(destLat) * math.cos(destLng - startLng)
-    brng = math.atan2(y, x)
-    brng = toDegrees(brng)
-    return (brng + 360) % 360
-
-def haversine_distance(lat1, lon1, lat2, lon2):
-    """
-    Calculate the great circle distance between two points
-    on the earth specified in decimal degrees
-    """
-    # Convert decimal degrees to radians
-    lat1, lon1, lat2, lon2 = map(math.radians, [lat1, lon1, lat2, lon2])
-
-    # Haversine formula
-    dlon = lon2 - lon1
-    dlat = lat2 - lat1
-    a = math.sin(dlat/2)**2 + math.cos(lat1) * math.cos(lat2) * math.sin(dlon/2)**2
-    c = 2 * math.asin(math.sqrt(a))
-    r = 6371  # Radius of earth in kilometers. Use 3956 for miles
-    return c * r
-
-def get_point_at_distance(lat1, lon1, d, bearing, reverse=False, R=6371):
-    """
-    lat: initial latitude, in degrees
-    lon: initial longitude, in degrees
-    d: target distance from initial
-    bearing: (true) heading in degrees
-    R: optional radius of sphere, defaults to mean radius of earth
-    reverse: In what directios should the point be on the line
-
-    Returns new lat/lon coordinate {d}km from initial, in degrees
-    """
-    lat1 = toRadians(lat1)
-    lon1 = toRadians(lon1)
-    bearing = toRadians(bearing)
-    if reverse:
-        bearing += math.pi
-    lat2 = math.asin(math.sin(lat1) * math.cos(d/R) + math.cos(lat1) * math.sin(d/R) * math.cos(bearing))
-    lon2 = lon1 + math.atan2(
-        math.sin(bearing) * math.sin(d/R) * math.cos(lat1),
-        math.cos(d/R) - math.sin(lat1) * math.sin(lat2)
-    )
-    return (toDegrees(lat2), toDegrees(lon2),)
 
 def save_polygon(id, p1, p2, p3, p4):
     if reverse_coordinates:
